Create a binary argument encoder to speed up argument encoding.
The introduction of the shell based argument encoder to pass the
command line to a long running opc command added a command processing
overhead of 4x. By creating a binary argument encoder we retain the
benefits of passing arguments in the environemnt, the command line is
very unlikely to be in the process list, while only slowing down the
CLI by 1.5x compared to arguments passed on the command line. There is
still a minor risk that the command line will appear in the process
list (since the arg encoder is short lived but must have the arguemnts
passed via the command line) but this existed with the old mechanism
as well. To help aleviate this the new binary clears its argv before
it encodes the data as JSON.

Fixes VRVDR-50650