opd: set PWD to / if $HOME does not exist
In runner.go:getExecCommand() $PWD is used to set the working
directory for the operational command being executed. If this
directory does not exist Cmd.Start() will fail.

The lack of a $HOME directory should not have the effect of
preventing any operational commands from being executed. This
condition can occur if a system enters a severe failed state where
the filesystem has gone readonly and a TACACS+ user logs in. In this
scenario we are unable to create the user's $HOME, and they are then
unable to eg. run the reboot op command.

VRVDR-52322