opd: Migrate to aaa plugin API v2
Update aaaAccount() to use the aaa.AAATask API and declare the
TaskAccounter interface, which is an abstration of and future
subset of the aaa.AAATask interface.

A response channel is added to the AcctReq struct allowing
TaskAccounter objects to be passed back from the auth layer.

This is required to support (TACACS+) command accounting start
records.