client: add doRpcOnConn()
Rename doRpc() and add an additional net.Conn argument which is used
to perform the request. The intention is to allow the net.Conn of the
Client to be wrapped by some callers in order to vary behaviour at
the connection read/write level.

For compatibility doRpc() is retained and delegates to doRpcOnConn().

VRVDR-53682