client: add intRequestOnConn()
Rename intRequest() and add an additional net.Conn argument which is
used to perform the request. The intention is to allow the net.Conn
of the Client to be wrapped by some callers in order to vary
behaviour at the connection read/write level.

Essentially this just exposes doRpcOnConn() to callers sending
requests which return an int eg. Client.Run().

For compatibility intRequest() is retained and delegates to
intRequestOnConn().

VRVDR-53682