opd: allow setting open terminal descriptor on Cmd
Typically runner.go:Run() will open a terminal descriptor for the
terminal specified by Termpath. Here we add a Term field to the
Cmd struct, allowing an open descriptor to be specified.

If Term is set, runner.go:Run() will use this instead of opening
a new descriptor, therefore allowing a command to be run with input
and outputs connected to an already opened terminal descriptor.

VRVDR-53682