opd: Update SrvConn to read SCM_RIGHTS message
Implement SrvConn:Read() which reads an SCM_RIGHTS ancilliary message,
containing a single file descriptor, from the underlying connection.

The descriptor, if any, is wrapped in a File object and stored, for
future use, by pointer on the SrvConn instance.

The expectation is that this descriptor will represent a user's PTY
which we can redirect command inputs and outputs to, as part of a Run
RPC.

VRVDR-53682