Pass PTY by descriptor from opc to opd
Currently opc specifies a PTY by path as an argument to the Run RPC.
opd then opens the PTY in order to read from/write to it, to pass
user input to the executed command and display its output to the user.

This presents an issue when opc is running in a different mount
namespace than opd, since the passed PTY path either does not exist
in opd's namespace or refers to an entirely different PTY (hence
taking input and displaying output to a different user).

Up to now we have avoided this by ensuring that all instances of opc
access the same devpts filesystem mounted on /dev/pts. However we
would like the ability to use isolated devpts filesystems for opc
instances which exist in a different namespace.

To address this opc now opens the PTY, hence in its own mount
namespace, and passes this file descriptor to opd, via an SCM_RIGHTS
message as part of the Run RPC request. The PTY path is no longer
passed as an argument to the Run RPC. opd then simply redirects
inputs and outputs of the executed command to this descriptor.

VRVDR-53682